Fix test search data processing

+++ app/src/main/java/com/utynote/components/search/data/geojson/entities/Geocoding.java
@@ -20,6 +20,6 @@ public class Geocoding {
     public Engine engine;
     @SerializedName("timestamp")
     @Expose
-    public Integer timestamp;
+    public long timestamp;
 
 }
